+++ test/test_query.cpp
@@ -7047,4 +7047,330 @@ TEST(Query_LinkCounts)
     CHECK_EQUAL(not_found, match);
 }
 
+TEST(Query_Link_Minimum)
+{
+    Group group;
+    TableRef table1 = group.add_table("table1");
+    table1->add_column(type_Int, "int", /* nullable */ true);
+    table1->add_column(type_Float, "float", /* nullable */ true);
+    table1->add_column(type_Double, "double", /* nullable */ true);
+
+    // table1
+    // 0: 789 789.0f 789.0
+    // 1: 456 456.0f 456.0
+    // 2: 123 123.0f 123.0
+    // 3: null null null
+
+    table1->add_empty_row();
+    table1->set_int(0, 0, 789);
+    table1->set_float(1, 0, 789.0f);
+    table1->set_double(2, 0, 789.0);
+    table1->add_empty_row();
+    table1->set_int(0, 1, 456);
+    table1->set_float(1, 1, 456.0f);
+    table1->set_double(2, 1, 456.0);
+    table1->add_empty_row();
+    table1->set_int(0, 2, 123);
+    table1->set_float(1, 2, 123.0f);
+    table1->set_double(2, 2, 123.0);
+    table1->add_empty_row();
+    table1->set_null(0, 3);
+    table1->set_null(1, 3);
+    table1->set_null(2, 3);
+
+    TableRef table2 = group.add_table("table2");
+    size_t col_linklist = table2->add_column_link(type_LinkList, "linklist", *table1);
+
+    // table2
+    // 0: { }
+    // 1: { 1 }
+    // 2: { 1, 2 }
+    // 3: { 1, 2, 3 }
+
+    table2->add_empty_row();
+
+    table2->add_empty_row();
+    LinkViewRef links = table2->get_linklist(col_linklist, 1);
+    links->add(1);
+
+    table2->add_empty_row();
+    links = table2->get_linklist(col_linklist, 2);
+    links->add(1);
+    links->add(2);
+
+    table2->add_empty_row();
+    links = table2->get_linklist(col_linklist, 3);
+    links->add(1);
+    links->add(2);
+    links->add(3);
+
+    Query q;
+    size_t match;
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).min() == 123;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).min() == 456;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).min() == null();
+    match = q.find();
+    CHECK_EQUAL(0, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).min() == 123.0f;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).min() == 456.0f;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).min() == 123.0;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).min() == 456.0;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+}
+
+TEST(Query_Link_MaximumSumAverage)
+{
+    Group group;
+    TableRef table1 = group.add_table("table1");
+    table1->add_column(type_Int, "int", /* nullable */ true);
+    table1->add_column(type_Float, "float", /* nullable */ true);
+    table1->add_column(type_Double, "double", /* nullable */ true);
+
+    // table1
+    // 0: 123 123.0f 123.0
+    // 1: 456 456.0f 456.0
+    // 2: 789 789.0f 789.0
+    // 3: null null null
+
+    table1->add_empty_row();
+    table1->set_int(0, 0, 123);
+    table1->set_float(1, 0, 123.0f);
+    table1->set_double(2, 0, 123.0);
+    table1->add_empty_row();
+    table1->set_int(0, 1, 456);
+    table1->set_float(1, 1, 456.0f);
+    table1->set_double(2, 1, 456.0);
+    table1->add_empty_row();
+    table1->set_int(0, 2, 789);
+    table1->set_float(1, 2, 789.0f);
+    table1->set_double(2, 2, 789.0);
+    table1->add_empty_row();
+    table1->set_null(0, 3);
+    table1->set_null(1, 3);
+    table1->set_null(2, 3);
+
+    TableRef table2 = group.add_table("table2");
+    size_t col_linklist = table2->add_column_link(type_LinkList, "linklist", *table1);
+
+    // table2
+    // 0: { }
+    // 1: { 1 }
+    // 2: { 1, 2 }
+    // 3: { 1, 2, 3 }
+
+    table2->add_empty_row();
+
+    table2->add_empty_row();
+    LinkViewRef links = table2->get_linklist(col_linklist, 1);
+    links->add(1);
+
+    table2->add_empty_row();
+    links = table2->get_linklist(col_linklist, 2);
+    links->add(1);
+    links->add(2);
+
+    table2->add_empty_row();
+    links = table2->get_linklist(col_linklist, 3);
+    links->add(1);
+    links->add(2);
+    links->add(3);
+
+    Query q;
+    size_t match;
+
+    // Maximum.
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).max() == 789;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).max() == 456;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).max() == null();
+    match = q.find();
+    CHECK_EQUAL(0, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).max() == 789.0f;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).max() == 456.0f;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).max() == 789.0;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).max() == 456.0;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    // Sum.
+    // Floating point results below may be inexact for some combination of architectures, compilers, and compiler flags.
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).sum() == 1245;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).sum() == 456;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).sum() == 1245.0f;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).sum() == 456.0f;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).sum() == 1245.0;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).sum() == 456.0;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    // Average.
+    // Floating point results below may be inexact for some combination of architectures, compilers, and compiler flags.
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).average() == 622.5;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).average() == 456;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Int>(0).average() == null();
+    match = q.find();
+    CHECK_EQUAL(0, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).average() == 622.5;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Float>(1).average() == 456.0f;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).average() == 622.5;
+    match = q.find();
+    CHECK_EQUAL(2, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(3, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+
+    q = table2->column<LinkList>(col_linklist).column<Double>(2).average() == 456.0;
+    match = q.find();
+    CHECK_EQUAL(1, match);
+    match = q.find(match + 1);
+    CHECK_EQUAL(not_found, match);
+}
+
 #endif // TEST_QUERY